+++ yamps/mps/tdvp.py
@@ -16,7 +16,7 @@ logger = logging.getLogger('yamps.tensor.tdvp')
 #################################
 
 
-def tdvp_OBC(psi, tmax, dt=1, H=False, M=False, env=None, cutoff_dE=1e-9, hermitian=True, fermionic=False, k=4, eigs_tol=1e-14, exp_tol=1e-14, dtype='complex128', bi_orth=True, NA=None, version='1site', opts_svd=None, optsK_svd=None):
+def tdvp_OBC(psi, tmax, dt=1, H=False, M=False, env=None, cutoff_dE=1e-9, hermitian=True, fermionic=False, k=4, eigs_tol=1e-14, exp_tol=1e-14, bi_orth=True, NA=None, version='1site', opts_svd=None, optsK_svd=None):
     # evolve with TDVP method, up to tmax and initial guess of the time step dt
     # opts - optional info for MPS truncation
     curr_t = 0
@@ -35,13 +35,13 @@ def tdvp_OBC(psi, tmax, dt=1, H=False, M=False, env=None, cutoff_dE=1e-9, hermit
                 'yamps.tdvp: Neither Hamiltonian nor Kraus operators defined.')
         else:
             if version == '2site':
-                env = tdvp_sweep_2site(psi=psi, H=H, M=M, dt=dt, env=env, dtype=dtype, hermitian=hermitian, fermionic=fermionic,
+                env = tdvp_sweep_2site(psi=psi, H=H, M=M, dt=dt, env=env, hermitian=hermitian, fermionic=fermionic,
                                        k=k, eigs_tol=eigs_tol, exp_tol=exp_tol, bi_orth=bi_orth, NA=NA, opts_svd=opts_svd, optsK_svd=optsK_svd)
             elif version == '2site_group':
-                env = tdvp_sweep_2site_group(psi=psi, H=H, M=M, dt=dt, env=env, dtype=dtype, hermitian=hermitian, fermionic=fermionic,
+                env = tdvp_sweep_2site_group(psi=psi, H=H, M=M, dt=dt, env=env, hermitian=hermitian, fermionic=fermionic,
                                              k=k, eigs_tol=eigs_tol, exp_tol=exp_tol, bi_orth=bi_orth, NA=NA, opts_svd=opts_svd, optsK_svd=optsK_svd)
             else:
-                env = tdvp_sweep_1site(psi=psi, H=H, M=M, dt=dt, env=env, dtype=dtype, hermitian=hermitian, fermionic=fermionic,
+                env = tdvp_sweep_1site(psi=psi, H=H, M=M, dt=dt, env=env, hermitian=hermitian, fermionic=fermionic,
                                        k=k, eigs_tol=eigs_tol, exp_tol=exp_tol, bi_orth=bi_orth, NA=NA, opts_svd=opts_svd, optsK_svd=optsK_svd)
         E = env.measure()
         dE = abs(E - E0)
@@ -51,7 +51,7 @@ def tdvp_OBC(psi, tmax, dt=1, H=False, M=False, env=None, cutoff_dE=1e-9, hermit
     return env, E, dE
 
 
-def tdvp_sweep_1site(psi, H=False, M=False, dt=1., env=None, dtype='complex128', hermitian=True, fermionic=False, k=4, eigs_tol=1e-14, exp_tol=1e-14, bi_orth=True, NA=None, opts_svd=None, optsK_svd=None):
+def tdvp_sweep_1site(psi, H=False, M=False, dt=1., env=None, hermitian=True, fermionic=False, k=4, eigs_tol=1e-14, exp_tol=1e-14, bi_orth=True, NA=None, opts_svd=None, optsK_svd=None):
     r"""
     Perform sweep with 1site-TDVP by applying exp(-i*dt*H) on initial vector. Note the convention for time step sign.
     Procedure performs exponantiation: psi(dt) = exp( dt * H )*psi(0). For Hamiltonian real time evolution forward in time: sign(dt)= -1j, for Hamiltonian imaginary time evolution forward in time: sign(dt)= -1.. For Hamiltonian real time evolution forward in time: sign(dt)= -1j, for Hamiltonian imaginary time evolution forward in time: sign(dt)= -1.
@@ -131,10 +131,10 @@ def tdvp_sweep_1site(psi, H=False, M=False, dt=1., env=None, dtype='complex128',
             init = psi.A[n]
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff1(v, n), init=[init], Bv=lambda v: env.Heff1(
-                    v, n, conj=True), dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol, k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                    v, n, conj=True), dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol, k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff1(v, n), init=[
-                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             init = init[0]
         # canonize and save
         psi.A[n] = init
@@ -147,10 +147,10 @@ def tdvp_sweep_1site(psi, H=False, M=False, dt=1., env=None, dtype='complex128',
             init = psi.A[psi.pC]
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff0(v, psi.pC), init=[init], Bv=lambda v: env.Heff0(
-                    v, psi.pC, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                    v, psi.pC, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff0(v, psi.pC), init=[
-                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             psi.A[psi.pC] = init[0]
             psi.absorb_central(towards=psi.g.last)
 
@@ -160,10 +160,10 @@ def tdvp_sweep_1site(psi, H=False, M=False, dt=1., env=None, dtype='complex128',
             init = psi.A[n]
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff1(v, n), init=[init], Bv=lambda v: env.Heff1(
-                    v, n, conj=True), dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                    v, n, conj=True), dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff1(v, n), init=[
-                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             init = init[0]
 
         if M:  # apply the Kraus operator
@@ -185,17 +185,17 @@ def tdvp_sweep_1site(psi, H=False, M=False, dt=1., env=None, dtype='complex128',
             init = psi.A[psi.pC]
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff0(v, psi.pC), init=[init], Bv=lambda v: env.Heff0(
-                    v, psi.pC, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                    v, psi.pC, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff0(v, psi.pC), init=[
-                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             psi.A[psi.pC] = init[0]
             psi.absorb_central(towards=psi.g.first)
 
     return env
 
 
-def tdvp_sweep_2site(psi, H=False, M=False, dt=1., env=None, dtype='complex128', hermitian=True, fermionic=False, k=4, eigs_tol=1e-14, exp_tol=1e-14, bi_orth=True, NA=None, opts_svd=None, optsK_svd=None):
+def tdvp_sweep_2site(psi, H=False, M=False, dt=1., env=None, hermitian=True, fermionic=False, k=4, eigs_tol=1e-14, exp_tol=1e-14, bi_orth=True, NA=None, opts_svd=None, optsK_svd=None):
     r"""
     Perform sweep with 2site-TDVP.
     Procedure performs exponantiation: psi(dt) = exp( dt * H )*psi(0). For Hamiltonian real time evolution forward in time: sign(dt)= -1j, for Hamiltonian imaginary time evolution forward in time: sign(dt)= -1.
@@ -279,10 +279,10 @@ def tdvp_sweep_2site(psi, H=False, M=False, dt=1., env=None, dtype='complex128',
             init = psi.A[n].dot(psi.A[n1], axes=(psi.right, psi.left))
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff2(v, n), Bv=lambda v: env.Heff2(v, n, conj=True), init=[
-                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff2(v, n), init=[
-                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             # split and save
             A1, S, A2 = init[0].split_svd(axes=(psi.left + psi.phys, tuple(
                 a + psi.right[0] - 1 for a in psi.phys + psi.right)), sU=-1, **opts_svd)
@@ -296,10 +296,10 @@ def tdvp_sweep_2site(psi, H=False, M=False, dt=1., env=None, dtype='complex128',
             init = psi.A[n1]
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff1(v, n1), init=[init], Bv=lambda v: env.Heff1(
-                    v, n1, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                    v, n1, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff1(v, n1), init=[
-                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             psi.A[n1] = init[0]
 
     for n in psi.g.sweep(to='first', df=1):
@@ -308,10 +308,10 @@ def tdvp_sweep_2site(psi, H=False, M=False, dt=1., env=None, dtype='complex128',
             init = psi.A[n]
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff1(v, n), init=[init], Bv=lambda v: env.Heff1(
-                    v, n, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                    v, n, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff1(v, n), init=[
-                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             psi.A[n] = init[0]
 
         if M:  # Apply the Kraus operator on n
@@ -332,10 +332,10 @@ def tdvp_sweep_2site(psi, H=False, M=False, dt=1., env=None, dtype='complex128',
             init = psi.A[n1].dot(psi.A[n], axes=(psi.right, psi.left))
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff2(v, n1), Bv=lambda v: env.Heff2(v, n1, conj=True), init=[
-                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff2(v, n1), init=[
-                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             # split and save
             A1, S, A2 = init[0].split_svd(axes=(psi.left + psi.phys, tuple(
                 a + psi.right[0] - 1 for a in psi.phys + psi.right)), sU=-1, **opts_svd)
@@ -349,7 +349,7 @@ def tdvp_sweep_2site(psi, H=False, M=False, dt=1., env=None, dtype='complex128',
     return env
 
 
-def tdvp_sweep_2site_group(psi, H=False, M=False, dt=1, env=None, dtype='complex128', hermitian=True, fermionic=False, k=4, eigs_tol=1e-14, exp_tol=1e-14, bi_orth=True, NA=None, opts_svd=None, optsK_svd=None):
+def tdvp_sweep_2site_group(psi, H=False, M=False, dt=1, env=None, hermitian=True, fermionic=False, k=4, eigs_tol=1e-14, exp_tol=1e-14, bi_orth=True, NA=None, opts_svd=None, optsK_svd=None):
     r"""
     Perform sweep with 2site-TDVP with grouping neigbouring sites.
     Procedure performs exponantiation: psi(dt) = exp( dt * H )*psi(0). For Hamiltonian real time evolution forward in time: sign(dt)= -1j, for Hamiltonian imaginary time evolution forward in time: sign(dt)= -1.
@@ -434,10 +434,10 @@ def tdvp_sweep_2site_group(psi, H=False, M=False, dt=1, env=None, dtype='complex
             init, leg_order = init.group_legs(axes=(1, 2), new_s=1)
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff2_group(v, n), Bv=lambda v: env.Heff2_group(v, n, conj=True), init=[
-                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff2_group(v, n), init=[
-                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             # split and save
             init = init[0].ungroup_leg(axis=1, leg_order=leg_order)
             A1, S, A2 = init.split_svd(axes=(psi.left + psi.phys, tuple(
@@ -452,10 +452,10 @@ def tdvp_sweep_2site_group(psi, H=False, M=False, dt=1, env=None, dtype='complex
             init = psi.A[n1]
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff1(v, n1), init=[init], Bv=lambda v: env.Heff1(
-                    v, n1, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                    v, n1, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff1(v, n1), init=[
-                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             psi.A[n1] = init[0]
 
     for n in psi.g.sweep(to='first', df=1):
@@ -464,10 +464,10 @@ def tdvp_sweep_2site_group(psi, H=False, M=False, dt=1, env=None, dtype='complex
             init = psi.A[n]
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff1(v, n), init=[init], Bv=lambda v: env.Heff1(
-                    v, n, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                    v, n, conj=True), dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff1(v, n), init=[
-                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=-dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             psi.A[n] = init[0]
 
         if M:  # Apply the Kraus operator on n
@@ -489,10 +489,10 @@ def tdvp_sweep_2site_group(psi, H=False, M=False, dt=1, env=None, dtype='complex
             init, leg_order = init.group_legs(axes=(1, 2), new_s=1)
             if not hermitian:
                 init = expmw(Av=lambda v: env.Heff2_group(v, n1), Bv=lambda v: env.Heff2_group(v, n1, conj=True), init=[
-                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth,  dtype=dtype, NA=NA)
+                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=False, bi_orth=bi_orth, NA=NA)
             else:
                 init = expmw(Av=lambda v: env.Heff2_group(v, n1), init=[
-                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, dtype=dtype, NA=NA)
+                             init], dt=+dt * .5, eigs_tol=eigs_tol, exp_tol=exp_tol,  k=k, hermitian=True, NA=NA)
             # split and save
             init = init[0].ungroup_leg(axis=1, leg_order=leg_order)
             A1, S, A2 = init.split_svd(axes=(psi.left + psi.phys, tuple(
